+++ resources/js/App.jsx
@@ -61,6 +61,12 @@ import AdminShopSettingsPage from './Pages/admin/AdminShopSettingsPage';
 import AdminAboutPage from './Pages/admin/AdminAboutPage';
 import AdminPersonalizedSolutionShowPage from './Pages/admin/AdminPersonalizedSolutionShowPage';
 import AdminPersonalizedSolutionEditPage from './Pages/admin/AdminPersonalizedSolutionEditPage';
+import AdminFaqsPage from './Pages/admin/AdminFaqsPage';
+import AdminFaqNewPage from './Pages/admin/AdminFaqNewPage';
+import AdminFaqEditPage from './Pages/admin/AdminFaqEditPage';
+import FaqPage from './Pages/FaqPage';
+import ForgotPasswordPage from './Pages/ForgotPasswordPage';
+import ResetPasswordPage from './Pages/ResetPasswordPage';
 import NotFoundPage from './Pages/NotFoundPage';
 import SessionExpiredPage from './Pages/SessionExpiredPage';
 
@@ -96,6 +102,9 @@ export default function App() {
                   <Route path="cart" element={<CartPage />} />
                   <Route path="login" element={<LoginPage />} />
                   <Route path="register" element={<RegisterPage />} />
+                  <Route path="faq" element={<FaqPage />} />
+                  <Route path="forgot-password" element={<ForgotPasswordPage />} />
+                  <Route path="reset-password" element={<ResetPasswordPage />} />
                   <Route path="checkout" element={<CheckoutPage />} />
                   <Route path="orders" element={<OrdersPage />} />
                   <Route path="purchases" element={<PurchasesPage />} />
@@ -126,6 +135,9 @@ export default function App() {
                   <Route path="products/new" element={<AdminProductNewPage />} />
                   <Route path="products/:id/edit" element={<AdminProductEditPage />} />
                   <Route path="products/:id" element={<AdminProductShowPage />} />
+                  <Route path="faqs/new" element={<AdminFaqNewPage />} />
+                  <Route path="faqs/:id/edit" element={<AdminFaqEditPage />} />
+                  <Route path="faqs" element={<AdminFaqsPage />} />
                   <Route path="features" element={<AdminFeaturesPage />} />
                   <Route path="features/new" element={<AdminFeatureNewPage />} />
                   <Route path="features/:id/edit" element={<AdminFeatureEditPage />} />