+++ resources/js/components/Layout.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Link, Outlet } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import Navbar from './Navbar';
+import EmailVerificationBanner from './EmailVerificationBanner';
 import CartWidget from './CartWidget';
 import Footer from './Footer';
 import ScrollToTop from './ScrollToTop';
@@ -30,6 +31,7 @@ export default function Layout() {
       <div className="drawer-content flex min-h-screen flex-col bg-base-200">
         <Navbar />
         <main className="container mx-auto min-w-0 max-w-full flex-1 px-4 py-6">
+          <EmailVerificationBanner />
           <Outlet />
         </main>
         <Footer />
@@ -57,6 +59,11 @@ export default function Layout() {
             <li>
               <Link to="/custom-solution" onClick={closeStorefrontDrawer}>
                 {t('shop.custom_solution')}
+              </Link>
+            </li>
+            <li>
+              <Link to="/faq" onClick={closeStorefrontDrawer}>
+                {t('shop.faq.nav')}
               </Link>
             </li>
             <li>