+++ resources/js/components/Navbar.jsx
@@ -161,6 +161,9 @@ export default function Navbar() {
             <Link to="/custom-solution" className="btn btn-ghost hidden sm:inline-flex shrink-0">
               {t('shop.custom_solution')}
             </Link>
+            <Link to="/faq" className="btn btn-ghost hidden md:inline-flex shrink-0">
+              {t('shop.faq.nav')}
+            </Link>
             <form onSubmit={handleSearch} className="join hidden lg:flex shrink-0 min-w-0">
               <input
                 type="search"